requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
bleach==3.1.5
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
dnspython==2.1.0
entrypoints==0.3
idna==2.10
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
packaging==20.4
pandocfilters==1.4.2
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.16.0
python-nmap==0.6.1
pytz==2020.1
pywin32==228
requests==2.25.1
six==1.14.0
sqlparse==0.3.1
Sublist3r==1.0
testpath==0.4.4
traitlets==4.3.3
urllib3==1.26.2
webencodings==0.5.1
win-unicode-console==0.5
wrapt==1.11.2
